+++ src/Resources/contao/languages/de/tl_nc_notification.php
@@ -1,0 +1,7 @@
+<?php
+
+$arrLang = &$GLOBALS['TL_LANG']['tl_nc_notification'];
+
+$arrLang['type'][\HeimrichHannot\ContaoNewsAlertBundle\EventListener\NewsPostedListener::NOTIFICATION_TYPE_NEWSALERT] = 'Newsalert';
+$arrLang['type'][\HeimrichHannot\ContaoNewsAlertBundle\EventListener\NewsPostedListener::NOTIFICATION_TYPE_NEW_ARTICLE] =
+    ['Neuer Artikel', 'Dieser Benachrichtigungstyp wird verwendet, um den Newsalert für einen neuen Artikel zu senden.'];